Lisp.Data

Description

This module contains all the Data Types necessary for the mini lisp interpretor.

Types

data Expr

Our Lisp expression

Constructors

Number Integer

Symbol String

Function Function Args

Special Function Args

List [Expr]

type Result = StateT Context HlError Expr

A state monad that holds a context and an evaluation result

type HlError = ErrorT String IO

A Transformer monad for handling error within the context

data Context

Context in which expressions will be evaluated

Constructors

Ctx SymbolTable (Maybe Context)