Lisp.Evaluation.Eval

Description

This module contains the Evaluation function for treating expressions.

Evaluation Function

eval :: Expr -> Result

Main Function for evaluating Lisp expression

Helper Functions

updateSymbol :: MonadState Context m => String -> Expr -> m ()

Update Symbol in the current Context

updateSymbolInParent :: MonadState Context m => [Char] -> Expr -> m ()

Update Symbol in the parent Context

popContext :: Context -> Context

Pop the parent Context within the current context

pushContext :: Context -> Context

Create a New context and hidding the current one within it