Lisp.Evaluation.StandardFunctions

Description

This module contains all the Standard Functions necessary for our lisp interpretor.

Initial Context

initialCtx :: Context

Our Initial symbol table

Standard Functions

define :: StateT Context HlError Expr

A function to modify the context

quote :: StateT Context HlError Expr

A function that return an argument without evaluate it

ifCm :: StateT Context HlError Expr

Conditionals

lambda :: StateT Context HlError Expr

Lambda Function

eq :: StateT Context HlError Expr

Equality between Expr

arithmetic :: (Integer -> Integer -> Integer) -> StateT Context HlError Expr

Main function for arithmetic operators

Helper Functions

binary :: (Integer -> Integer -> Integer) -> [Expr] -> Result

Take the Arithmetic Function, the Arguments and returning the result

getSymbol :: String -> Result

Get a symbol associated value

getSymbols :: [String] -> StateT Context HlError [Expr]

Map the getSymbol function over a list as arguments